requirements.txt
braintree==3.59.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
class-registry==2.1.2
colorama==0.4.3
cryptography==3.1.1
defusedxml==0.7.0rc1
Django==2.2
django-jazzmin==2.4.1
django-simple-coupons==0.5
django-contrib-comments==1.9.2
django-crispy-forms==1.9.2
django-csvimport==2.16
django-easy-pdf==0.1.1
django-extensions==3.0.9
django-filter==2.4.0
django-qr-code==1.3.1
django-qrcode==0.3
django-rest-authtoken==2.1.3
django-rest-registration==0.5.6
django-social-share==2.0.0
djangorestframework==3.12.0
filters==1.3.2
filters-django==1.0.5
html5lib==1.1
idna==2.10
Markdown==3.2.2
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
PyPDF2==1.26.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
qrcode==6.1
regex==2020.9.27
reportlab==3.5.51
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
typing==3.7.4.3
urllib3==1.25.10
webencodings==0.5.1
Werkzeug==1.0.1
xhtml2pdf==0.2.4
